+++ PineappleTimer WatchKit Extension/ExtensionDelegate.swift
@@ -28,6 +28,7 @@ class ExtensionDelegate: NSObject, WKExtensionDelegate, UNUserNotificationCenter
 
     func applicationDidBecomeActive() {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
+        clearAllDeliveredNotifications()
     }
 
     func applicationWillResignActive() {
@@ -77,6 +78,10 @@ func requestNotificationPermissions() {
             return
         }
     }
+}
+
+private func clearAllDeliveredNotifications() {
+    UNUserNotificationCenter.current().removeAllDeliveredNotifications()
 }
 
 private func scheduleLocalNotifications() {